sqlite: bail out of `explain` if the program loops
I had a simple `COUNT(*) GROUP BY` query in my app, and it caused
the compilation to hang indefinitely. The steps sqlite generated
to execute the query contained a loop (it did `Goto index 1` at the end).

This patch is not a solution to the problem of infinite loops but at
least it will bail out instead of hanging.